fix: hub structural fixes — lock serialization, promotion tracking, counter safety (#449)
Moved hub write lock to SyncManager (#457, #459):
- HubWriteLock and acquire_hub_lock moved from shared_writer/core.rs
  to sync/cache.rs where they belong — the lock protects SyncManager's
  hub cache, not SharedWriter specifically.
- SyncManager::acquire_lock() is the single entry point.
- fetch() now acquires the lock before modifying the working directory,
  preventing daemon/CLI races where a daemon fetch could corrupt a
  CLI's pending commit.
- upgrade_to_v2() now acquires the lock before changing layout,
  preventing agents from writing V1 files while migration runs.
- write_commit_push() uses self.sync.acquire_lock() instead of
  directly calling the lock function.

Promotion UUID tracking (#451):
- record_promoted_uuids() error now propagates instead of being
  swallowed. If tracking fails, promotion returns Err so the caller
  knows the state is incomplete. Prevents re-promotion of already-
  promoted UUIDs on next sync.

Counter collisions (#450) and compaction race (#453):
- Closed as safe by design after code audit. Counter claims re-read
  from disk inside the write lock on every retry. Materialization
  runs inside the compaction lock.
